Add workaround for Nokogiri script/style CDATA bug.
Nokogiri 1.4.2 and higher have a fragment parsing bug (issue #315 in
Nokogiri's GitHub issue tracker) that causes "</body></html>" to be
appended to the CDATA inside unterminated script and style elements.

This commit adds a special transformer that identifies affected elements
and removes the spurious text when Nokogiri 1.4.2 or higher is being
used.