Create whitespace nodes in a way that's less likely to make Nokogiri segfault.
On an Arch Linux maching running Ruby 1.9.2, Nokogiri 1.4.4, and libxml
2.7.8, Nokogiri would sometimes segfault here. I'm not sure what the
exact set of circumstances is that triggers it, but it seems to be
at least partially environmental. I could reproduce the problem
consistently given certain inputs when running in one app, but the same
inputs didn't cause a problem in irb or in another app, and I couldn't
reproduce the problem in the original app on another machine.