Add data-uri image support in the relaxed config
This allows people to provide content that has `<img src="data:image/gif...">` in it. Previously, `<img src="data:image/gif...">` would result in an empty image tag being returned. With this change, the image will actually show up.

Based on preliminary research, I believe this is a safe default and makes sense to be in the core relaxed config.