fix: Strip null bytes before passing input to Nokogumbo
This fixes crashes in recent versions due to a failed assertion in
Gumbo.

The specific trigger seems to be a null byte in the value of an unquoted
attribute that contains non-markup quote characters:

    <img src=javascript:alert('\0')>

Several of Sanitize's unit tests test input like this, which didn't fail
until relatively recently when something must have changed in Gumbo.