Remove optimization to get back the previous behaviour of transformers
Before [1] transformers could modify the DOM, and transformers or
filtering that ran after would use the modified DOM. The change in [1]
meant that the node name was cached between the running of all
transformers so even if one transformer modified it subsequent
transformers would be given the previous node name.

The added test case fails without the code change because the 'strong'
tag is stripped out after the transformer is run since it thinks it is
still a 'b'.

[1] -
https://github.com/rgrove/sanitize/commit/caa558a25886638427d3005030260af8518612ea