Bring back memory optimization of caching lowercased node name
In 4.6.2, #175 significantly cut down on memory usage by caching the lowercase
version of the node name. However, as reported in #177, this broke certain
transformers that modified the node name since the name was cached. We can
bring back this optimization by updating the node name only if it has been
changed.

Closes #184