feat: Remove the content of whitelisted iframes
The HTML spec says iframes shouldn't have content, but when they do,
it's parsed as text and is serialized verbatim without being escaped,
which is unsafe because legacy browsers may still render it and execute
`<script>` content. So the safe and correct thing to do is to always
remove iframe content.

Prior to Nokogumbo 2.0 iframe content was escaped when serialized so it
wasn't an issue, but Nokogumbo 2.0 has a spec-compliant serializer so
this is now important.