Stop using the term "whitelist"
The term "whitelist" has been replaced with "allowlist" throughout
Sanitize's source and documentation.

While the etymology of "whitelist" may not be explicitly racist in
origin or intent, there are inherent racial connotations in the
implication that white is good and black (as in "blacklist") is not.

This is a change I should have made long ago, and I apologize for not
making it sooner.