Forcibly escape content in "unescaped text" elements inside math or svg namespaces
This fixes an edge case in which the contents of an "unescaped text"
element (such as `<noembed>` or `<xmp>`) were not properly escaped if
that element was allowlisted and was also inside an allowlisted `<math>`
or `<svg>` element.

The only way to encounter this situation was to ignore multiple warnings
in the readme and create a custom config that allowlisted all the
elements involved, including `<math>` or `<svg>`.

Please let this be a reminder that Sanitize cannot safely sanitize
MathML or SVG content and does not support this use case. The default
configs don't allow MathML or SVG elements, and allowlisting MathML or
SVG elements in a custom config may create a security vulnerability in
your application.

Documentation has been updated to add more warnings and to make the
existing warnings about this more prominent.